Modularize and optimize parsing of arguments - Rename `__main__.py` to `runner.py` - Extract arg parsing related functions to `argparser.py` - Move `main()` to `main.py`
